Выполнено задание "Модуляция".
