+++ js/const.js
@@ -1,0 +1,21 @@
+const ADS_COUNT = 10;
+
+const LOCATION = {
+  n1: 35.65, n2: 35.70, e1: 139.70, e2: 139.80, floatNumber: 5,
+};
+
+const REALTY_TYPES = ['palace', 'flat', 'house', 'bungalow', 'hotel',];
+
+const CHECK_VALUES = ['12:00', '13:00', '14:00',];
+
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner',];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+export {
+  ADS_COUNT, LOCATION, REALTY_TYPES, CHECK_VALUES, FEATURES, PHOTOS,
+};
